upload single image file when post tweet

+++ app/src/main/java/crazysheep/io/nina/holder/timeline/DraftHolder.java
@@ -19,7 +19,9 @@ import butterknife.ButterKnife;
 import crazysheep.io.nina.R;
 import crazysheep.io.nina.bean.PostTweetBean;
 import crazysheep.io.nina.prefs.UserPrefs;
+import crazysheep.io.nina.utils.DebugHelper;
 import crazysheep.io.nina.utils.DialogUtils;
+import crazysheep.io.nina.utils.Utils;
 import de.hdodenhof.circleimageview.CircleImageView;
 
 /**
@@ -107,7 +109,11 @@ public class DraftHolder extends BaseHolder<PostTweetBean> implements View.OnCli
                         super.onClick(dialog);
 
                         // delete draft from database and update UI
-                        mPostTweetBean.delete();
+                        // TODO check why post tweet bean id is null int database
+                        DebugHelper.log("remove draft, id: " + mPostTweetBean.getId());
+                        if(!Utils.isNull(mPostTweetBean.getId())) {
+                            mPostTweetBean.delete();
+                        }
                         EventBus.getDefault().post(new EventRemoveDraft(mPostTweetBean));
                     }
                 },
